backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow==10.1.0
djoser==2.2.0
webcolors==1.13
